backend/requirements.txt
alembic==1.14.1
fastapi==0.115.12
pydantic-settings==2.8.1
pymysql==1.1.1
pytest==8.3.5
python-dotenv==1.0.1
sqlalchemy==2.0.40
strawberry-graphql[fastapi]==0.262.3
uvicorn[standard]==0.34.0
